refactor(harness): split prompt context rendering
Ultraworked with [omo](https://github.com/code-yeongyu/oh-my-openagent)
