+++ packages/backend/src/harness/prompt-design-system.ts
@@ -1,0 +1,76 @@
+import type { buildSessionContext } from "../services/context";
+import { readOptional } from "./prompt-file-reader";
+
+type SessionContext = NonNullable<
+  Awaited<ReturnType<typeof buildSessionContext>>
+>;
+type DesignSystem = NonNullable<SessionContext["designSystem"]>;
+
+export const MAX_SKILL_CHARS = 5000;
+const MAX_TOKENS_CSS_LINES = 150;
+const MAX_README_LINES = 120;
+
+export async function appendDesignSystemContext(
+  lines: string[],
+  designSystem: DesignSystem,
+  contextMode: "compact" | "full",
+): Promise<void> {
+  lines.push("## Design system");
+  lines.push(`- name: ${designSystem.name}`);
+  lines.push(`- directory: ${designSystem.dir_path}`);
+  if (designSystem.skill_md_path) {
+    lines.push(`- skill: ${designSystem.skill_md_path}`);
+  }
+  if (designSystem.tokens_css_path) {
+    lines.push(`- tokens: ${designSystem.tokens_css_path}`);
+  }
+  if (designSystem.readme_md_path) {
+    lines.push(`- readme: ${designSystem.readme_md_path}`);
+  }
+  lines.push("");
+
+  if (contextMode === "compact") {
+    lines.push("### Compact design-system handling");
+    lines.push(
+      "- Use the design-system paths above as source of truth. Read SKILL.md, tokens, or README only when exact brand rules or token names are needed for this request.",
+    );
+    lines.push(
+      "- Prefer targeted Grep/Read ranges over loading full design-system files. Reuse existing CSS variables instead of inventing new palettes or type stacks.",
+    );
+    lines.push("");
+    return;
+  }
+
+  if (designSystem.skill_md_path) {
+    const content = await readOptional(designSystem.skill_md_path);
+    if (content) {
+      lines.push("### SKILL.md");
+      lines.push("```markdown");
+      lines.push(content.slice(0, MAX_SKILL_CHARS));
+      lines.push("```");
+      lines.push("");
+    }
+  }
+  if (designSystem.tokens_css_path) {
+    const content = await readOptional(designSystem.tokens_css_path);
+    if (content) {
+      lines.push("### colors_and_type.css (excerpt)");
+      lines.push("```css");
+      lines.push(
+        content.split("\n").slice(0, MAX_TOKENS_CSS_LINES).join("\n"),
+      );
+      lines.push("```");
+      lines.push("");
+    }
+  }
+  if (designSystem.readme_md_path) {
+    const content = await readOptional(designSystem.readme_md_path);
+    if (content) {
+      lines.push("### README.md (excerpt)");
+      lines.push("```markdown");
+      lines.push(content.split("\n").slice(0, MAX_README_LINES).join("\n"));
+      lines.push("```");
+      lines.push("");
+    }
+  }
+}
