+++ packages/backend/src/harness/prompt-file-reader.ts
@@ -1,0 +1,9 @@
+import { readFile } from "node:fs/promises";
+
+export async function readOptional(filePath: string): Promise<string | null> {
+  try {
+    return await readFile(filePath, "utf8");
+  } catch {
+    return null;
+  }
+}
